remove consoles
