+++ src/Components/Loader/reducer.ts
@@ -1,5 +1,8 @@
-import { loadingDefaultState } from "./types";
 import { createSlice } from "@reduxjs/toolkit";
+
+export type LoaderState = boolean;
+
+export const loadingDefaultState: LoaderState = true;
 
 export const loaderSlice = createSlice({
   name: "loader",
