--- src/Components/Loader/types.ts
@@ -1,3 +1,0 @@
-export type LoaderState = boolean;
-
-export const loadingDefaultState: LoaderState = true;
