--- src/Components/NameForm/types.ts
@@ -1,22 +1,0 @@
-enum errorTypes {
-  loginError,
-  logoutError,
-}
-
-type errorLog = Array<keyof typeof errorTypes>;
-
-interface AuthState {
-  name: string;
-  isAuth: boolean;
-  errorLog: errorLog;
-}
-
-export const authDefaultState: AuthState = {
-  name: "",
-  isAuth: false,
-  errorLog: [],
-};
-
-export type payloadType = {
-  name: string;
-};
